c: fix format-string warning about using %d with unsigned int
clang 19.1.4 (on Alpine 3.21) reports:

    find_aes_key.c:1545:60: error: format specifies type 'int' but the
    argument has type 'unsigned int' [-Werror,-Wformat]
     1544 |                     printf("[%#" PRIx64 "..%#" PRIx64 "] Found AES-%d decryption key: ",
          |                                                                    ~~
          |                                                                    %u
     1545 |                            addr, addr + expanded_key_size, key_len * 8);
          |                                                            ^~~~~~~~~~~
    1 error generated.

    libc_memory.c:63:12: error: format specifies type 'long' but the
    argument has type 'uintptr_t' (aka 'unsigned long') [-Werror,-Wformat]
       62 |     printf("Difference: %" PRIdPTR " = %u + %" PRIdPTR "\n",
          |                         ~~~~~~~~~~
       63 |            addr_high - addr_low, allocsize, addr_high - addr_low - allocsize);
          |            ^~~~~~~~~~~~~~~~~~~~
    libc_memory.c:63:45: error: format specifies type 'long' but the
    argument has type 'uintptr_t' (aka 'unsigned long') [-Werror,-Wformat]
       62 |     printf("Difference: %" PRIdPTR " = %u + %" PRIdPTR "\n",
          |                                             ~~~~~~~~~~
       63 |            addr_high - addr_low, allocsize, addr_high - addr_low - allocsize);
          |                                             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
    libc_memory.c:172:50: error: format specifies type 'unsigned int'
    but the argument has type 'int' [-Werror,-Wformat]
      172 |     printf("Allocated %u 0x%x-byte block in:\n", NUM_ALLOCATED_BLOCKS, allocsize);
          |                       ~~                         ^~~~~~~~~~~~~~~~~~~~
          |                       %d
    libc_memory.c:152:30: note: expanded from macro 'NUM_ALLOCATED_BLOCKS'
      152 | #define NUM_ALLOCATED_BLOCKS 5
          |                              ^
    3 errors generated.

    setjmp.c:704:38: error: format specifies type 'long' but the argument has type 'unsigned long' [-Werror,-Wformat]
      704 |         printf("  [%2ld: @%#05lx] ", reg_idx, reg_idx * (unsigned int)sizeof(uintptr_t));
          |                    ~~~~              ^~~~~~~
          |                    %2lu
    1 error generated.

    x86-tables.c:190:82: error: format specifies type 'unsigned int' but the argument has type 'int' [-Werror,-Wformat]
      190 |     printf("CPU %2d GDT @%p, size %u (%u 8-byte entries)\n", cpu, address, size, (size + 1) / 8);
          |                                       ~~                                         ^~~~~~~~~~~~~~
          |                                       %d
    x86-tables.c:208:84: error: format specifies type 'unsigned int' but the argument has type 'int' [-Werror,-Wformat]
      208 |     printf("CPU %2d IDT @%p, size %u (256 %u-byte vectors)\n", cpu, address, size, (size + 1) / 256);
          |                                           ~~                                       ^~~~~~~~~~~~~~~~
          |                                           %d
    2 errors generated.

    ...

Fix all these issues.

+++ c/x86-segments.c
@@ -307,7 +307,7 @@ static void print_gdt_limits(void)
 
     memcpy(&gdt_size, gdt_descriptor, 2);
     if (gdt_size) {
-        printf("GDT limits and access rights (%u entries):\n", (gdt_size + 1) / 8);
+        printf("GDT limits and access rights (%u entries):\n", ((unsigned int)gdt_size + 1) / 8);
     } else {
         /* Try 256 entries */
         printf("GDT limits and access rights (? entries):\n");
